feat: refactor theme management by creating a dedicated context and hook
