+++ src/hooks/use-theme.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react"
+import { ThemeProviderContext } from "./theme-provider-context"
+
+export const useTheme = () => {
+	const context = useContext(ThemeProviderContext)
+
+	if (context === undefined) throw new Error("useTheme must be used within a ThemeProvider")
+
+	return context
+}
